Special optimizer where all optimal objective values are known.
