+++ src/main/java/org/clafer/compiler/ClaferCompiler.java
@@ -259,7 +259,7 @@ public class ClaferCompiler {
                     ? new EquivalentParetoSolver(new ClaferSingleObjectiveOptimizer(solver, solution, maximizes[0], objectiveVars[0]))
                     : new EquivalentParetoSolver(new ClaferMultiObjectiveOptimizerGIA(solver, solution, maximizes, objectiveVars));
         } catch (UnsatisfiableException e) {
-            return new ClaferUnsatOptimizer();
+            return new ClaferNoObjectiveOptimizer(new ClaferSolver());
         }
     }
